Rennes_metro_project/api/requirements.txt
requests
mysql-connector-python

